+++ examples/express/webpack.config.js
@@ -1,55 +1,66 @@
 import { resolve, basename, dirname, posix } from "path";
-import externals from "webpack-node-externals";
-import { TogglePointInjection } from "@asos/web-toggle-point-webpack/plugins";
-import MiniCssExtractPlugin from "mini-css-extract-plugin";
 import { fileURLToPath } from "url";
+import externals from "webpack-node-externals";
+import { TogglePointInjectionPlugin } from "@asos/web-toggle-point-webpack/plugins";
+import MiniCssExtractPlugin from "mini-css-extract-plugin";
+import staticLoadStrategyFactory from "@asos/web-toggle-point-webpack/moduleLoadStrategyFactories/staticLoadStrategyFactory";
+import lazyComponentLoadStrategyFactory from "@asos/web-toggle-point-react-pointcuts/lazyComponentLoadStrategyFactory";
 
 const configPointCutConfig = {
   name: "configuration variants",
   variantGlobs: ["./src/routes/config/__variants__/*/*/*.jsx"],
-  togglePointModule: "/src/routes/config/togglePoint.js"
+  togglePointModuleSpecifier: "/src/routes/config/togglePoint.js",
+  loadStrategy: lazyComponentLoadStrategyFactory(),
 };
 
 const common = {
   mode: "production",
+  devtool: "source-map",
+  experiments: {
+    outputModule: true,
+  },
+  output: {
+    module: true,
+  },
   module: {
     rules: [
       {
         test: /\.jsx?$/,
         exclude: /node_modules/,
         use: {
-          loader: "babel-loader"
-        }
+          loader: "babel-loader",
+        },
       },
       {
         test: /\.css$/,
-        use: [MiniCssExtractPlugin.loader, "css-loader"]
-      }
-    ]
-  }
+        use: [MiniCssExtractPlugin.loader, "css-loader"],
+      },
+    ],
+  },
+  plugins: [new MiniCssExtractPlugin()],
 };
 
 const config = [
   {
     entry: "./src/index.js",
     target: "node",
-    output: {
-      path: resolve(dirname(fileURLToPath(import.meta.url)), "bin"),
-      filename: "server.cjs",
-      clean: true,
-      chunkFormat: "module"
-    },
-    externals: [externals()],
     ...common,
+    output: {
+      ...common.output,
+      path: resolve(dirname(fileURLToPath(import.meta.url)), "bin"),
+      filename: "server.mjs",
+      clean: true,
+    },
+    externals: [externals({ importType: "module" })],
     plugins: [
-      new MiniCssExtractPlugin(),
-      new TogglePointInjection({
+      ...common.plugins,
+      new TogglePointInjectionPlugin({
         pointCuts: [
           configPointCutConfig,
           {
             name: "animal apis by version",
             variantGlobs: [
-              "./src/routes/animals/api/**/v{1..9}*([[:digit:]])/*.js"
+              "./src/routes/animals/api/**/v{1..9}*([[:digit:]])/*.js",
             ],
             joinPointResolver: (variantPath) =>
               posix.resolve(
@@ -57,25 +68,38 @@ const config = [
                 ...Array(3).fill(".."),
                 basename(variantPath)
               ),
-            togglePointModule: "/src/routes/animals/togglePoint.js"
-          }
-        ]
-      })
-    ]
+            togglePointModuleSpecifier: "/src/routes/animals/togglePoint.js",
+            loadStrategy: staticLoadStrategyFactory(),
+          },
+        ],
+      }),
+    ],
   },
   {
     entry: "./src/routes/config/client.js",
     target: "web",
+    ...common,
     output: {
+      ...common.output,
       path: resolve(dirname(fileURLToPath(import.meta.url)), "public"),
-      filename: "main.js"
+      filename: "main.mjs",
     },
     plugins: [
-      new MiniCssExtractPlugin(),
-      new TogglePointInjection({ pointCuts: [configPointCutConfig] })
+      ...common.plugins,
+      new TogglePointInjectionPlugin({ pointCuts: [configPointCutConfig] }),
     ],
-    ...common
-  }
+    module: {
+      ...common.module,
+      rules: [
+        {
+          test: /\.js$/,
+          enforce: "pre",
+          use: ["source-map-loader"],
+        },
+        ...common.module.rules,
+      ],
+    },
+  },
 ];
 
 export default config;