+++ examples/next/src/app/fixtures/content-management/__pointCutConfig.js
@@ -1,0 +1,10 @@
+export default [
+  {
+    name: "content management",
+    togglePointModuleSpecifier:
+      "/src/app/fixtures/content-management/withToggledHook",
+    variantGlobs: [
+      "./src/app/fixtures/content-management/**/__variants__/*/*/use!(*.spec).ts"
+    ]
+  }
+];