+++ examples/serve/src/fixtures/translation/__pointCutConfig.js
@@ -1,8 +1,11 @@
 import joinPointResolver from "../../joinPointResolver.js";
+// eslint-disable-next-line import/no-unresolved -- https://github.com/import-js/eslint-plugin-import/issues/1810
+import loadStrategyFactory from "@asos/web-toggle-point-webpack/moduleLoadStrategyFactories/staticLoadStrategyFactory";
 
 export default {
   name: "translation",
-  togglePointModule: "/src/fixtures/translation/__togglePoint.js",
+  togglePointModuleSpecifier: "/src/fixtures/translation/__togglePoint.js",
   variantGlobs: ["./src/fixtures/translation/languages/*/*.json"],
-  joinPointResolver
+  joinPointResolver,
+  loadStrategy: loadStrategyFactory()
 };