+++ packages/webpack/src/plugins/togglePointInjection/index.js
@@ -1,11 +1,12 @@
 import processPointCuts from "./processPointCuts/index.js";
 import Logger from "./logger.js";
-import { win32, posix } from "path";
+import { sep, posix } from "path";
 import { PLUGIN_NAME } from "./constants.js";
 import resolveJoinPoints from "./resolveJoinPoints/index.js";
 import setupSchemeModules from "./setupSchemeModules/index.js";
 import { validate } from "schema-utils";
 import schema from "./schema.json";
+import fillDefaultOptionalValues from "./fillDefaultOptionalValues.js";
 
 /**
  * Toggle Point Injection Plugin
@@ -17,23 +18,30 @@ class TogglePointInjection {
    * Create a {@link https://webpack.js.org/concepts/plugins/|Plugin} that injects toggle points into a Webpack build
    * @param {object} options plugin options
    * @param {object[]} options.pointCuts toggle point point cut configuration, with target toggle point code as advice. The first matching point cut will be used
-   * @param {string} options.pointCuts[].name name to describe the nature of the point cut, for clarity in logs and dev tools etc.
-   * @param {string} options.pointCuts[].togglePointModule path, from root of the compilation, of where the toggle point sits. Or a resolvable node_module.
-   * @param {string[]} [options.pointCuts[].variantGlobs=[.\/**\/__variants__/*\/*\/!(*.test).{js,jsx,ts,tsx}]] {@link https://en.wikipedia.org/wiki/Glob_(programming)|Globs} to identified variant modules.  The plugin uses {@link https://github.com/mrmlnc/fast-glob|fast-glob} under the hood, so supports any glob that it does.
-   * @param {function} [options.pointCuts[].joinPointResolver=(variantPath) => path.posix.resolve(variantPath, "../../../..", path.basename(variantPath))] A function that takes the path to a variant module and returns a join point / base module.  N.B. This is executed at build-time, so cannot use run-time context. It should use posix path segments, so on Windows be sure to use path.posix.resolve.
-   * @param {string} [options.pointCuts[].toggleHandler] Path to a toggle handler that unpicks a {@link https://webpack.js.org/api/module-methods/#requirecontext|require.context} containing potential variants, passing that plus a joint point module to a toggle point function. If not provided, the plugin will use a default handler that processes folder names into a tree held in a Map. Leaf nodes of the tree are the variant modules.
+   * @param {string} options.pointCuts[].name name to describe the nature of the point cut, for clarity in logs and dev tools etc
+   * @param {string} options.pointCuts[].togglePointModuleSpecifier a {@link https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules#importing_features_into_your_script|module specifier} pointing to the toggle point module
+   * @param {string} [options.pointCuts[].variantGlobs=['.\/**\/__variants__/*\/*\/!(*.test).{js,jsx,ts,tsx}']] {@link https://en.wikipedia.org/wiki/Glob_(programming)|Glob} to identified variant modules.  The plugin uses {@link https://github.com/mrmlnc/fast-glob|fast-glob} under the hood, so supports any glob that it does
+   * @param {function} [options.pointCuts[].joinPointResolver=(variantPath) => path.posix.resolve(variantPath, "../../../..", path.basename(variantPath))] A function that takes the path to a variant module and returns a join point / base module.  N.B. This is executed at build-time, so cannot use run-time context. It should use posix path segments, so on Windows be sure to use path.posix.resolve
+   * @param {string} [options.pointCuts[].toggleHandlerFactoryModuleSpecifier='@asos/web-toggle-point-webpack/toggleHandlerFactories/pathSegment'] a {@link https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules#importing_features_into_your_script|module specifier} pointing to a toggle handler factory, that takes a toggle point, and returns a handler that unpicks a Map of relative paths to potential variants, passing that plus a joint point. If not provided, the plugin will use a default handler that processes folder names into a tree held in a Map. The join point and leaf nodes of the tree are modules in a form defined by the loading strategy
+   * @param {module:web-toggle-point-webpack.loadStrategy} [options.pointCuts[].loadStrategy] a module load strategy. default is one created by "@asos/web-toggle-point-webpack/moduleLoadStrategyFactories/deferredRequireLoadStrategyFactory"
    * @param {function} [options.webpackNormalModule] A function that returns the Webpack NormalModule class.  This is required for Next.js, as it does not expose the NormalModule class directly
    * @returns {external:Webpack.WebpackPluginInstance} WebpackPluginInstance
-   * @example <caption>N.B. forward slashes are escaped in the glob, due to JSDoc shortcomings, but in reality should be un-escaped</caption>
+   * @example <caption>N.B. forward slashes + asterisk are escaped in the examples, due to JSDoc shortcomings, but in reality should be un-escaped</caption>
+   * import loadStrategyFactory from "@asos/web-toggle-point-react-pointcuts/lazyComponentLoadStrategyFactory";
    * const plugin = new TogglePointInjection({
    *   pointCuts: [
    *     {
-   *       togglePointModule: "/withToggledHook",
+   *       togglePointModuleSpecifier: "/withToggledHook",
    *       variantGlobs: ["./**\/__variants__/*\/*\/use!(*.test).{ts,tsx}"]
    *     },
    *     {
-   *       togglePointModule: "/withTogglePoint",
-   *       variantGlobs: ["./**\/__variants__/*\/*\/!(use*|*.test).{ts,tsx}"]
+   *       togglePointModuleSpecifier: "/withTogglePoint",
+   *       variantGlobs: ["./**\/__variants__/*\/*\/!(use*|*.test).tsx"],
+   *       loadStrategy: loadStrategyFactory({
+   *          importCodeGeneratorFactoryOptions: {
+   *             webpackMagicComment: "/* webpackPrefetch *\/"
+   *          }
+   *       })
    *     }
    *   ]
    * });
@@ -44,17 +52,13 @@ class TogglePointInjection {
    */
   constructor(options) {
     validate(schema, options, { name: PLUGIN_NAME, baseDataPath: "options" });
-    this.options = {
-      webpackNormalModule: async () =>
-        (await import("webpack")).default.NormalModule,
-      ...options
-    };
+    this.options = fillDefaultOptionalValues(options);
   }
 
   apply(compiler) {
     let NormalModule, joinPointFiles, warnings, appRoot;
     compiler.hooks.beforeCompile.tapPromise(PLUGIN_NAME, async () => {
-      appRoot = compiler.context.replaceAll(win32.sep, posix.sep);
+      appRoot = compiler.context.replaceAll(sep, posix.sep);
       ({ joinPointFiles, warnings } = await processPointCuts({
         appRoot,
         fileSystem: compiler.inputFileSystem,