+++ packages/webpack/src/plugins/togglePointInjection/setupSchemeModules/generatePointCut.js
@@ -1,13 +1,19 @@
-const generatePointCut = ({ pointCuts, path }) => {
-  const pointCutName = path.slice(1);
-  const {
-    togglePointModule,
-    toggleHandler = "@asos/web-toggle-point-webpack/pathSegmentToggleHandler"
-  } = pointCuts.find(({ name }) => name === pointCutName);
+import { posix, sep } from "path";
 
-  return `import togglePoint from "${togglePointModule}";
-import handler from "${toggleHandler}";
-export default (rest) => handler({ togglePoint, ...rest });`;
+const generatePointCut = ({ pointCuts, joinPointPath }) => {
+  const pointCutName = joinPointPath.slice(1);
+  const {
+    togglePointModuleSpecifier,
+    toggleHandlerFactoryModuleSpecifier,
+    loadStrategy: { adapterModuleSpecifier }
+  } = pointCuts.find(({ name }) => name === pointCutName);
+  return `import togglePoint from "${togglePointModuleSpecifier}";
+import handlerFactory from "${toggleHandlerFactoryModuleSpecifier}";
+import * as namespace from "${adapterModuleSpecifier.replaceAll(sep, posix.sep)}";
+const identity = (module) => module;
+const { pack:_pack = identity, unpack:_unpack = identity } = namespace;
+const handler = handlerFactory({ togglePoint, pack: _pack, unpack: _unpack });
+export default handler;`;
 };
 
 export default generatePointCut;