+++ test/automation/docs/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - ??
+
+### Added
+
+- added `:ui` scripts, for those that want to watch their automation
+
+### Changed
+
+- updated Playwright to 1.52.0
+
 ## [0.1.2] - 2024-12-18
 
 ### Fixed